requirements.txt
django ~= 3.1
djangorestframework ~= 3.12
psycopg2-binary
drf-yasg ~= 1.20
Khayyam ~= 3.0
djangorestframework-simplejwt ~= 4.4
django-filter ~= 2.3
Pillow ~= 8.1
python-decouple ~= 3.3
PyJWT==1.7.1
django-tinymce ~= 3.2
bs4 ~= 0.0.1
django-json-widget ~= 1.1.0